safari-600.7-branch/LayoutTests/platform/ios-sim/storage/indexeddb/intversion-two-opens-no-versions-expected.txt

CONSOLE MESSAGE: line 14: TypeError: undefined is not an object (evaluating 'request.onsuccess = deleteSuccess')
Test that only the first open call gets an upgradeneeded

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-two-opens-no-versions.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onsuccess = deleteSuccess')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
